entries by categories without sort after fixing bugs caused by eslint
