+++ static/src/js/components/tournament/entries_and_seeds/entries_and_seeds_index.js
@@ -29,6 +29,7 @@ const entriesTableStyle = {
   backgroundColor: 'white',
 };
 
+
 function sortBy(data, prop, isAsc) {
   isAsc = isAsc === undefined ? true : isAsc;
 
@@ -67,6 +68,8 @@ export default class EntriesAndSeeds extends React.Component {
 
   filterEntriesByName = (entry, name) => entry.tournament_category === name;
   componentDidMount() {
+    console.log('componentDidMount:: loading...');
+
     const { TournamentStore, CategoryStore } = this.props.stores;
 
     // Bring current tournament
@@ -75,6 +78,7 @@ export default class EntriesAndSeeds extends React.Component {
       : ' ';
     this.setState({ tournament });
 
+    console.log('tournament', tournament);
     const self = this;
     // Bring categories by the tournament id
     CategoryStore.CategoriesByTournament(tournament.id).then((storedCategories) => {
@@ -84,16 +88,18 @@ export default class EntriesAndSeeds extends React.Component {
 
   componentDidUpdate(prevProps, prevState) {
     if (this.state.categories !== prevState.categories) {
+      console.log('category: ', this.state.categories);
       const self = this;
-
-      if (this.state.categories) {
-        this.state.categories.map((category, index) => {
+      this.state.categories
+        ? this.state.categories.map((category, index) => {
           const { EntriesStore } = self.props.stores;
 
-          if (category && EntriesStore) {
-            EntriesStore.fetchEntriesByCategory(category.id).then((storedPlayers) => {
+          category && EntriesStore
+            ? EntriesStore.fetchEntriesByCategory(category.id).then((storedPlayers) => {
               entrieDoc.push(mobx.toJS(storedPlayers));
-              self.setState({ entries: entrieDoc }, () => {});
+              self.setState({ entries: entrieDoc }, () => {
+                console.log('entries: ', this.state.entries);
+              });
               // let rawData = entrieDoc[category.id];
               //
               // console.log('componentDidMount::rawData', [].concat(rawData));
@@ -102,10 +108,10 @@ export default class EntriesAndSeeds extends React.Component {
               //     , 'rank'/* param To Sort By */
               //     , false, /* sort ascending */
               // );
-            });
-          }
-        });
-      }
+            })
+            : console.log('loading');
+        })
+        : console.log('loading categories: ', this.state.categories);
     }
   }
 
@@ -129,7 +135,12 @@ export default class EntriesAndSeeds extends React.Component {
       </TableRow>
     );
 
-    const entriesTable = num => (
+    const entriesTable = (num) => {
+      {
+        console.log('num: ', num);
+        console.log('entries[num]: ', this.state.entries[num]);
+      }
+      return (
         <div>
           <Table style={entriesTableStyle}>
             <TableHeader displaySelectAll={false}>
@@ -150,7 +161,8 @@ export default class EntriesAndSeeds extends React.Component {
             </TableBody>
           </Table>
         </div>
-    );
+      );
+    };
     return (
       <div>
         <Tabs onChange={this.handleChange} value={this.state.slideIndex}>
